* Move BinaryParser into utils
